+++ packages/shadow-objects/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to [@spearwolf/shadow-objects](https://github.com/spearwolf/
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+
+- sharpen the `EntityApi` type definitions
+
 ## [0.26.4] - 2026-01-15
 
 - fix return type definitions for `provideContext()` and `provideGlobalContext()`